fix: Precision erros

+++ Sources/PeakSwift/Utils/Helper/LinearFilter.swift
@@ -30,8 +30,7 @@ enum LinearFilter {
     }
     
     private static func createImpulseResponse(samplingFrequency: Double, c: Double) -> [Double] {
-       // Sometime precision erros that may lead to deviations to neurokit
-       let denominator = samplingFrequency * c
+       let denominator = Int(samplingFrequency * c)
        let impulseResponeTemp: [Double] = Array(repeating: 1, count: Int(denominator))
         
         let impulseRespone = impulseResponeTemp.map { x in
